requirements.txt
google-api-python-client>=2.149.0
google-auth>=2.35.0
anthropic>=0.42.0
requests>=2.32.0
pyyaml>=6.0.2
youtube-transcript-api>=1.0.0
notebooklm-py>=0.3.0
